+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/warning/view/NotEnoughSpaceView.kt
@@ -1,7 +1,6 @@
 package co.electriccoin.zcash.ui.screen.warning.view
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -9,15 +8,15 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.design.component.Body
 import co.electriccoin.zcash.ui.design.component.GradientSurface
@@ -38,34 +37,40 @@ private fun NotEnoughSpacePreview() {
     }
 }
 
-// TODO [#883]: NotEnoughSpace screen has dark theme hardcoded
-// TODO [#883]: https://github.com/Electric-Coin-Company/zashi-android/issues/883
-
 @Composable
 fun NotEnoughSpaceView(
     storageSpaceRequiredGigabytes: Int,
     spaceRequiredToContinueMegabytes: Int
 ) {
-    @Suppress("MagicNumber")
-    val backgroundColor = Color(0xFF1A233A)
     Column(
         Modifier
-            .background(backgroundColor)
             .fillMaxSize()
-            .padding(32.dp),
+            .padding(ZcashTheme.dimens.screenHorizontalSpacingRegular)
+            .verticalScroll(
+                rememberScrollState()
+            ),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Image(painterResource(id = R.drawable.not_enough_space), "", Modifier.fillMaxWidth())
-        Spacer(Modifier.height(32.dp))
-        Header(text = stringResource(id = R.string.not_enough_space_title), color = Color.White)
-        Spacer(Modifier.height(32.dp))
+        Image(
+            painter = painterResource(id = R.drawable.not_enough_space),
+            contentDescription = stringResource(id = R.string.not_enough_space_logo_content_description),
+            modifier = Modifier.fillMaxWidth()
+        )
+
+        Spacer(Modifier.height(ZcashTheme.dimens.spacingXlarge))
+
+        Header(text = stringResource(id = R.string.not_enough_space_title))
+
+        Spacer(Modifier.height(ZcashTheme.dimens.spacingXlarge))
+
         Body(
             text = stringResource(id = R.string.not_enough_space_description, storageSpaceRequiredGigabytes),
-            textAlign = TextAlign.Center,
-            color = Color.White
+            textAlign = TextAlign.Center
         )
-        Spacer(Modifier.height(64.dp))
+
+        Spacer(Modifier.height(ZcashTheme.dimens.spacingHuge))
+
         Small(
             text = stringResource(id = R.string.space_required_to_continue, spaceRequiredToContinueMegabytes),
             textAlign = TextAlign.Center,